Allow `batch` to be a tuple in `batch_to_device`
While evaluating using the `LabelAccuracyEvaluator`, the batch are not dictionaries, but rather a tuple containing a dictionary as first value, and a tensor as second value. This causes `batch_to_device(...)` to fail with error `TypeError: tuple indices must be integers or slices, not list`.

I fixed `batch_to_device` to be more tolerating in what types it accepts. Possibly `LabelAccuracyEvaluator` could use a refactor to avoid this scheme, but I think the solution I propose has the advantage of always outputting a clear error message.