Make Cleaning Of Strings In NoDuplicatesDataLoader Optional
With very large datasets with a high percentage of duplicates you will som times need to sys.intern the strings to save memory by only having each uniq string in memory once.

Currently the NoDuplicatesDataLoader calls strip and lower on all the strings as part of the dedupe process which then means a new string is added to memory.

To this end we propose adding an option to NoDuplicatesDataLoader that will allow you to skip the text cleaning portion of the DataLoader to ensure that multiples of the same string are not pulled into memory.

Default behavior will be to clean the strings to ensure backwards compatibility.